+++ projects/np-ui-lib/src/lib/np-data-grid/services/np-file.service.ts
@@ -1,22 +1,24 @@
+import { DatePipe } from '@angular/common';
 import { Injectable } from '@angular/core';
 import { Column } from '../models/column.model';
+import { DataTypes } from '../models/constants';
 
 @Injectable()
 export class NpFileService {
-    downloadCSVFile(data: any[], columns: any[], fileName?: string) {
-        const CSVData = this.createCSVData(data, columns);
+
+    constructor(private datePipe: DatePipe) { }
+
+    downloadCSVFile(data: any[], columns: any[], dateFormat: string, fileName?: string) {
+        const CSVData = this.createCSVData(data, columns, dateFormat);
         return this.downloadFile(CSVData, this.normalizeFileName(fileName, 'csv'));
     }
 
     private downloadFile(content: string, fileName: string): void {
         const dataURI = this.generateDataURI(content);
-
         const anchor = document.createElement('a');
         anchor.href = dataURI;
-
         anchor.download = fileName;
         anchor.setAttribute('style', 'visibility:hidden');
-
         document.body.appendChild(anchor);
         anchor.click();
         document.body.removeChild(anchor);
@@ -26,24 +28,27 @@ export class NpFileService {
         return 'data:text/csv;charset=utf-8,\uFEFF' + encodeURIComponent(content);
     }
 
-    private createCSVData(data: any[], columnList: Column[]) {
+    private createCSVData(data: any[], columnList: Column[], dateFormat: string) {
         const headersLabels = [];
-        const headers = [];
         for (const element of columnList) {
             if (element.dataField === undefined) {
                 continue;
             }
             headersLabels.push(element.caption);
-            headers.push(element.dataField);
         }
         const head = headersLabels.join(',') + '\r\n';
         const rows = [];
         for (const element of data) {
             const cols = [];
-            for (const column of headers) {
-                if (element[column] !== undefined || element[column] != null) {
-                    cols.push(`"${element[column].toString().replace(/\"/g, '""')}"`);
-                } else {
+            for (const column of columnList) {
+                if (element[column.dataField] !== undefined && element[column.dataField] !== null) {
+                    if (column.dataType === DataTypes.Date) {
+                        cols.push(this.datePipe.transform(element[column.dataField], dateFormat));
+                    } else {
+                        cols.push(`"${element[column.dataField].toString().replace(/\"/g, '""')}"`);
+                    }
+                }
+                else {
                     cols.push('');
                 }
             }