+++ src/app/how-to-add/how-to-add.component.ts
@@ -5,9 +5,15 @@ import { Component, OnInit } from '@angular/core';
   templateUrl: './how-to-add.component.html'
 })
 export class HowToAddComponent implements OnInit {
+
+  importCss = `"styles": [
+    "node_modules/@angular/cdk/overlay-prebuilt.css",
+    "node_modules/np-ui-lib/style.css",
+    ...
+],`;
+
   constructor() { }
 
   ngOnInit(): void {
   }
-
 }