+++ src/app/np-grid-layout-demo/np-grid-layout-demo.module.ts
@@ -3,13 +3,15 @@ import { CommonModule } from '@angular/common';
 
 import { NpGridLayoutDemoRoutingModule } from './np-grid-layout-demo-routing.module';
 import { NpGridLayoutDemoComponent } from './np-grid-layout-demo.component';
+import { NpCardComponent, NpCardModule } from 'np-ui-lib';
 
 
 @NgModule({
   declarations: [NpGridLayoutDemoComponent],
   imports: [
     CommonModule,
-    NpGridLayoutDemoRoutingModule
+    NpGridLayoutDemoRoutingModule,
+    NpCardModule
   ]
 })
 export class NpGridLayoutDemoModule { }