+++ src/app/np-virtual-scroll-demo/np-virtual-scroll-demo.component.ts
@@ -1,5 +1,5 @@
-import { Component, OnInit } from '@angular/core';
-import { BehaviorSubject } from 'rxjs';
+import { Component, OnInit, ViewChild } from '@angular/core';
+import { NpVirtualScrollComponent } from 'np-ui-lib';
 
 @Component({
   selector: 'app-np-virtual-scroll-demo',
@@ -8,7 +8,7 @@ import { BehaviorSubject } from 'rxjs';
 export class NpVirtualScrollDemoComponent implements OnInit {
 
   importText = 'import { NpVirtualScrollModule } from \'np-ui-lib\';';
-  htmlText = `<np-virtual-scroll [data]="data" [pageSize]="pageSize" [itemHeight]="50" [template]="itemTemplate" (loadData)="loadData($event)">
+  htmlText = `<np-virtual-scroll [items]="data" [pageSize]="pageSize" [itemHeight]="50" [template]="itemTemplate" (loadData)="loadData($event)">
 </np-virtual-scroll>
 
 <ng-template #itemTemplate let-item="item">
@@ -23,12 +23,13 @@ loadData(event) {
     Array.prototype.splice.apply(this.data, [...[event.first, event.rows], ...fetchedData]);
     this.data = [...this.data];
   }, 2000);
-}
-`;
+}`;
 
   pageSize: number;
   data: any[];
   allData: any[];
+
+  @ViewChild('virtualScroll') virtualScroll: NpVirtualScrollComponent;
 
   constructor() {
     this.pageSize = 5;
@@ -48,6 +49,18 @@ loadData(event) {
 
       this.data = [...this.data];
     }, 2000);
+  }
+
+  scrollTo11Item() {
+    this.virtualScroll.scrollToIndex(10);
+  }
+
+  scrollOffsetDown() {
+    this.virtualScroll.scrollToOffset(1000);
+  }
+
+  scrollOffsetUp() {
+    this.virtualScroll.scrollToOffset(-1000);
   }
 
 }